wall rendition; bot factory

+++ core/src/main/java/org/ken22/input/settings/BotSettings.java
@@ -42,7 +42,7 @@ public class BotSettings {
 
     // Newton-Raphson
     @JsonProperty
-    public double nrMaxIterations;
+    public int  nrMaxIterations;
     @JsonProperty
     public double nrTolerance;
     @JsonProperty
